Fix bug in model checker clock vector logic
*** Important bug fix ***
This commit resolves a subtle bug in the model checking logic for
detecting reversible races. The DPOR paper of Flanagan and Godefroid
uses clock vectors to determine in O(1) time whether two transitions in
the transition stack (or trace) t_i and t_j are in a "happens-before"
relationship (calculating the clocks requires traversing the stack
however). The original pseudocode reads

i -->_S j <----> i <= C(j)(proc(S_i))

where S = {1, 2, ..., N} is the sequence of transitions, i and j are in
[1, N], S_i is the `i`th transition, proc(S_i) is the thread id of the
`i`th transition, and `C(j)` is the value of the `j`th clock vector.

The question arises to where `C(j)` is stored. The pseudocode notes that
`C([p := cv2, |S'| := cv2]`. That is, the clock vector `C[l]` where
`l` is the subtrace including transition `l`, is contained in the `l +
1`th state stack position, NOT the `lth`.