Add initial implementation of greedy relinearization
When exploring new traces, DPOR states that any exploration forward is
correct. When there are multiple enabled threads at any given point
which haven't been previously explored, DPOR can choose between any one
of them. Histotrically, McMini chose to schedule the thread with the
lowest id first before others. Recently, round-robin scheduling was
introduced that allowed McMini instead to fairly schedule threads for as
long as possible. The goal was to enable McMini to find shallower bugs
more quickly.

The challenge with round robin scheduling is that it produces traces
that are challenging to analyze. This is because a trace will
constantly be switching between threads to schedule making it difficult
for a user to see what's going on. The solution is to relinearize the
traces. The goal is to produce a new trace that obeys the same happens
before relation but which is more readable. There are two approaches to
doing this:

1. Choose the linearization that always schedules groups of threads for
   as long as possible (greedy)
2. Produce a linearization that minimizes the number of context switches
   a person must go through when analyzing the trace.

The latter is much more challenging (and possibly NP-hard although this
is only a hunch). The former is very easy.