Fix subtle dependency issue in ddt
The double dispatch table has two means of handling functions: either a
single interface exists that applies to all other instances, or else a
specific function exists overriding a specific dependency pairing. The
goal is to enable some operations to apply a uniform rule regardless of
the other transition we're comparing to (e.g. a thread start is
independent of everything, and instead of needing to specify N functions
one for each other kind of operation we can instead specify it once).
There was an edge case in which comparison with a type which has
registered _other_ specific handlers resulted in the global handlers not
being called. This happened because we didn't subsequently check if a
global function existed if the pairing failed.