Print relinearized traces when given specific flags
McMini now will print different relinearizations of traces when given
either the `-relin` or `-orelin` flags. The former corresponds to the
greedy relinearization approach while the latter corresponds to the
optimal reordering based on the SCIP solver. McMini must be compiled
with SCIP; otherwise, optimal reordering will fallback to the greedy
approach